api pronta de verbos http

+++ springfield/src/main/java/com/springfield/cidade/cidadao/CidadaoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CidadaoRepository extends JpaRepository<Cidadao, Long> {
+public interface CidadaoRepository extends JpaRepository<Cidadao, Integer> {
 }
